%% File Name: TAM-defunct.Rd
%% File Version: 0.07
%% File Last Change: 2017-06-22 17:41:52

\name{TAM-defunct}
\alias{TAM-defunct}
\alias{tam.jml2}



\title{Defunct \pkg{TAM} Functions}

\description{
These functions have been removed or replaced in the \pkg{tam.jml2}
package. 
}

\usage{	
tam.jml2(...)
}

\arguments{
	\item{\dots}{Arguments to be passed.}
}

\details{
The \code{tam.jml2} is included as the default in 
\code{\link{tam.jml}}.	
}

% The \code{rasch.pml2} function has been superseeded by
% \code{\link{rasch.pml3}}.

% The \code{yen.q3} function has been replaced by
% \code{\link{Q3}}.


% \examples{
% #############################################################################
% }
