\name{TAM-package}
\alias{TAM-package}
\alias{TAM}
\docType{package}
\title{
Test Analysis Modules
}
\description{
  The package includes marginal and joint maximum likelihood
  estimation of uni- and multidimensional item response models 
  (Rasch,  2PL, Generalized Partial Credit, Rating Scale, Multi Facets),
  fit statistics, standard error estimation, as 
  well as plausible value imputation and weighted likelihood
  estimation of ability.
}
\details{
\tabular{ll}{
Package: \tab \pkg{TAM}\cr
Type: \tab Package\cr
Version: \tab 0.6\cr
Publication Year: \tab 2013\cr
License: \tab GPL (>=2.0)\cr
URL: \tab \url{https://sites.google.com/site/irttam1/},  \cr
\tab \url{http://www.edmeasurementsurveys.com/TAM/Tutorials/} 
}

See \url{http://www.edmeasurementsurveys.com/TAM/Tutorials/} for 
tutorials of the \pkg{TAM} package.
}
\author{
Thomas Kiefer [aut], Alexander Robitzsch [aut, cre], Margaret Wu [aut]
}
\references{
Adams, R. J., Wilson, M., & Wu, M. (1997).
Multilevel item response models: An approach to errors in
variables regression. \emph{Journal of Educational and Behavioral 
Statistics}, \bold{22}, 47-76.

Adams, R. J., & Wu, M. L. (2007). The mixed-coefficients multinomial logit model.
A generalized form of the Rasch model. In M. von Davier & C. H. Carstensen (Eds.):
\emph{Multivariate and mixture distribution Rasch models: Extensions and applications}
(pp. 55-76). New York: Springer.
}
\keyword{ package }
%\seealso{
%%~~ Optional links to other man pages, e.g. ~~
%%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%% ## ~~ simple examples of the most important functions ~~
%}
