%% File Name: TAM-package.Rd
%% File Version: 2.653

\name{TAM-package}
\alias{TAM-package}
\alias{TAM}
\docType{package}

\title{
\packageTitle{TAM}
}

\description{
\packageDescription{TAM}
}

\author{
\packageAuthor{TAM}

Maintainer: \packageMaintainer{TAM}
}




\details{
See \url{http://www.edmeasurementsurveys.com/TAM/Tutorials/} for
tutorials of the \pkg{TAM} package.
}

\references{
Adams, R. J., Wilson, M., & Wang, W. C. (1997). The multidimensional random coefficients
multinomial logit model. \emph{Applied Psychological Measurement, 21}(1), 1-23.
\doi{10.1177/0146621697211001}

Adams, R. J., Wilson, M., & Wu, M. (1997).
Multilevel item response models: An approach to errors in
variables regression. \emph{Journal of Educational and Behavioral
Statistics, 22}(1), 47-76. \doi{10.3102/10769986022001047}

Adams, R. J., & Wu, M. L. (2007). The mixed-coefficients multinomial logit model.
A generalized form of the Rasch model. In M. von Davier & C. H. Carstensen (Eds.):
\emph{Multivariate and mixture distribution Rasch models: Extensions and applications}
(pp. 55-76). New York: Springer.
\doi{10.1007/978-0-387-49839-3_4}

Formann, A. K. (1982). Linear logistic latent class analysis.
\emph{Biometrical Journal, 24}(2), 171-190.
\doi{10.1002/bimj.4710240209}

Formann, A. K. (1992). Linear logistic latent class analysis for polytomous data.
\emph{Journal of the American Statistical Association, 87}(418), 476-486.
\doi{10.1080/01621459.1992.10475229}
}


\keyword{package}


%\seealso{
%%~~ Optional links to other man pages, e.g. ~~
%%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%% ## ~~ simple examples of the most important functions ~~
%}
