\name{TAM-package}
\alias{TAM-package}
\alias{TAM}
\docType{package}
\title{
Test Analysis Modules
}
\description{
  Includes marginal and joint maximum likelihood
  estimation of uni- and multidimensional item response models 
  (Rasch, 2PL, 3PL, Generalized Partial Credit, Multi Facets,
  Nominal Item Response, Structured Latent Class Analysis,
  Mixture Distribution IRT Models, Located Latent Class Models). 
  Latent regression models and plausible value imputation are 
  also supported.
}
\details{
\tabular{ll}{
Package: \tab \pkg{TAM}\cr
Type: \tab Package\cr
Version: \tab 1.12\cr
Publication Year: \tab 2015\cr
License: \tab GPL (>=2.0)\cr
URL: \tab \url{http://www.edmeasurementsurveys.com/TAM/Tutorials/}\cr
 \tab \url{https://sites.google.com/site/alexanderrobitzsch/software/} \cr
}
% \tab \url{https://www.bifie.at/} \cr 

See \url{http://www.edmeasurementsurveys.com/TAM/Tutorials/} for 
tutorials of the \pkg{TAM} package.

}
\author{
Thomas Kiefer [aut], Alexander Robitzsch [aut, cre], Margaret Wu [aut]

Maintainer: Alexander Robitzsch <a.robitzsch@bifie.at>
}
\references{
Adams, R. J., Wilson, M., & Wu, M. (1997).
Multilevel item response models: An approach to errors in
variables regression. \emph{Journal of Educational and Behavioral 
Statistics}, \bold{22}, 47-76.

Adams, R. J., & Wu, M. L. (2007). The mixed-coefficients multinomial logit model.
A generalized form of the Rasch model. In M. von Davier & C. H. Carstensen (Eds.):
\emph{Multivariate and mixture distribution Rasch models: Extensions and applications}
(pp. 55-76). New York: Springer.

Formann, A. K. (1992). Linear logistic latent class analysis for polytomous data. 
\emph{Journal of the American Statistical Association}, \bold{87}, 476-486.
}
\keyword{package}
%\seealso{
%%~~ Optional links to other man pages, e.g. ~~
%%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%% ## ~~ simple examples of the most important functions ~~
%}
