\name{TAM-utilities}
\alias{TAM-utilities}
\alias{tam_packageinfo}
\alias{tam_rsessinfo}
\alias{tam_print_call}


\title{Utility Functions in \pkg{TAM}}

\description{
Utility functions in \pkg{TAM}. 
}

\usage{
## information about used package version
tam_packageinfo(pack)
## call statement in a string format	
tam_print_call(CALL)
## information about R session
tam_rsessinfo()
}

\arguments{
\item{pack}{An \R package}
\item{CALL}{An \R call}
}

%\details{
%xxx
%%}


% \examples{
% #############################################################################
% }
